Do the -webkit-transform:translate in js.
The -webkit-transform:translate(-4px, -4px) was added to move the hint
labels a little bit from the hinted links so that they do not cover to
much of the link text. But in this way the labels where sometime only
partial visible if the hinted element is near to the top or left.

This patch removes the css translate and calculates the offset direct
in the hints.js. So we can now ensure that the hint labels are not
placed behind the top or left of the viewport.