Use g_string_erase instead of memmove.
The memmove moved queued keys toward the beginning and left clutter and
unwanted stuff at the end which might cause issues in future. So now
g_string_erase is used which strips chars from the beginning of the
string. This is what the memmove() intended but did not make obvious.