config.mk: link against webkit-gtk 4.1
The webkit-gtk project has released a new minor version bump of their
API with version 4.1. The difference between 4.0 and 4.1 is that the
former links against libsoup 2.x, whereas the latter links against
libsoup 3.0 and newer. As libsoup has introduced backwards-incompatible
changes in 3.0 it will cause the process to hard-crash if a program is
linked against both libraries at the same time.

Now that we have ported vimb to be compatible with libsoup 3.0 we can
start to link against the recent webkit-gtk 4.1.